Rephrase Justification for Inclusion of Small Multiples
@FransWi: Could you add one sentence with the reason why you take this
as the state-of-the-art?

Old: We use this conventional technique as the state-of-the-art to
compare our innovations against.

New: This conventional techniques serves as a benchmark to compare our
innovations against.

The reason is that I haven't seen other techniques used in print. But
then again I'm not aware of many Lexis-Surfaces of compositional data.
Coming to think of it there isn't a state-of-the-art for showing
compositional data on the Lexis-Surface. Accordingly I rephrased the
sentence.