Clarify Table-Lookup Criterion
@FransWi: Is a criterion of good visualization the presence of table
look-up? Is yes, say it explicitly.

Response: Not necessarily. Sometimes pattern-perception is all that it
needed. However, when scientists discuss a graph they tend to refer to
numerical data points ("...look at this outlier at 16.2", "There is an
increase of about 2.5 per year"...). Table look up allows for numerical
judgements. I added this statement to the definition.