Add Justification for Use of Colour
@FransWi: This section would benefit from an introduction on why colors
are important in display of compositional data.
For instance: " Colours are used to map the composition of data." I like
the basic terminology used in the grammer of graphics, but you do should
not use it of you do not like it.

I also suggest you tel the reader why we have a section on colours:
"iThis section ntroduces colour terminology and the concepts of colour
composition and colour spaces. These concepts are used in the
description of visualization techniques."

Response: The different visualizations in this paper exploit different
features of colour in order to solve the problem of showing composite
data on a plane. I rewrote the section starting off with an explanation
on why colour is useful in our case.

I agree that Wilkinsons colour terminology comes more natural (he uses
Brightness, Hue and Saturation instead of Lightness, Hue and Chroma) but
I stick with the terms used in the CIELhc colour space we use in the
paper. Otherwise we would use two different terminologies in the same
paper.