Add Paragraph on Colour Schemes to "About Colour" Section
@AnnaKl: I do not know what these are. Is it necessarry to know? If yes,
please explain. If not, I would rather leave it out.
@FransWi: What makes a palette qualitative rather than ?? It may not be
clear to the reader what is meant by qualitative palette of different
hues.

Response: Ok. I just realized that knowing this stuff actually is
important for the article. Also, given that 2 of the 4 visualizations in
this paper draw heavily from Brewers colour typology I decided to add a
section on colour scales/schemes to the "About Colour" section.