Separate LaTeX Styling Commands from Content
I extracted all the LaTeX styling and commands into a separate .sty
file. This file can be included as a package.

I have separate .sty files for my personal preferences and for the
demands of Demograhic Research. Changing the .sty file to use in the
main document I can change the look of the document.